refact(useFilter): hoisting of results array after filter
